+++ src/platform/__init__.py
@@ -1,7 +1,7 @@
 """
 Platform services: API, job queue, data lake, distributed runners.
 """
-from .api_server import run_api_server
+from .api_server import create_api_server, run_api_server
 from .backtest_task import run_backtest_job
 from .data_lake import DataLake
 from .distributed import run_distributed_backtests
@@ -9,6 +9,7 @@ from .job_queue import JobQueue, JobStore
 from .orchestrator import run_workflow
 
 __all__ = [
+    "create_api_server",
     "run_api_server",
     "run_backtest_job",
     "DataLake",